Validacion de nombre repetido en crear curso y editar curso corregidos Referenciado a #63
