+++ src/main/java/com/sofkau/academicsystembackend/usecases/course/UpdateCourseUseCase.java
@@ -1,12 +1,13 @@
 package com.sofkau.academicsystembackend.usecases.course;
 
 import com.sofkau.academicsystembackend.models.course.CourseDTO;
+import com.sofkau.academicsystembackend.repositories.CourseRepository;
+import com.sofkau.academicsystembackend.utils.ValidationsCourse;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 import reactor.core.publisher.Mono;
 
 import java.util.Objects;
-import java.util.stream.IntStream;
 
 @Service
 @Validated
@@ -14,10 +15,16 @@ public class UpdateCourseUseCase implements SaveCourse{
 
     private final CreateCourseUseCase createCourseUseCase;
     private final GetCourseByIdUseCase getCourseByIdUseCase;
+    private final CourseRepository courseRepository;
+    private final MapperUtilsCourse mapperUtilsCourse;
+    private final ValidationsCourse validationsCourse;
 
-    public UpdateCourseUseCase(CreateCourseUseCase createCourseUseCase, GetCourseByIdUseCase getCourseByIdUseCase) {
+    public UpdateCourseUseCase(CreateCourseUseCase createCourseUseCase, GetCourseByIdUseCase getCourseByIdUseCase, CourseRepository courseRepository, MapperUtilsCourse mapperUtilsCourse) {
         this.createCourseUseCase = createCourseUseCase;
         this.getCourseByIdUseCase = getCourseByIdUseCase;
+        this.courseRepository = courseRepository;
+        this.mapperUtilsCourse = mapperUtilsCourse;
+        this.validationsCourse = new ValidationsCourse(this.courseRepository, this.mapperUtilsCourse);
     }
 
     @Override
@@ -27,6 +34,10 @@ public class UpdateCourseUseCase implements SaveCourse{
         return getCourseByIdUseCase.apply(courseDTO.getId())
                 .flatMap(courseFoundDTO -> {
 
+                    if(courseDTO.getName().equalsIgnoreCase(courseFoundDTO.getName())) {
+                        return updateCourse(courseDTO);
+                    }
+
                     if(!Objects.isNull(courseFoundDTO.getName())){
                         return createCourseUseCase.apply(courseDTO);
                     }
@@ -34,4 +45,15 @@ public class UpdateCourseUseCase implements SaveCourse{
                     return Mono.error(new IllegalArgumentException("El curso no existe"));
                 });
     }
+
+    public Mono<CourseDTO> updateCourse(CourseDTO courseDTO) {
+        Mono<CourseDTO> error = validationsCourse.isErrorSave(courseDTO);
+        if (error != null) return error;
+        if (validationsCourse.isValidSaveForRules(validationsCourse.countRulesPresent(courseDTO)).get(0)) {
+            return courseRepository.save(mapperUtilsCourse.mapperToCourse().apply(courseDTO))
+                    .map(mapperUtilsCourse.mapperEntityToCourse());
+        }
+        return Mono.error(new IllegalArgumentException("Las reglas solo deben ser tres por categoria"));
+    }
+
 }
